+++ visualization/app/codeCharta/ui/viewCube/viewCube.mouseEvents.service.ts
@@ -1,4 +1,3 @@
-import { IRootScopeService } from "angular"
 import { hierarchy } from "d3-hierarchy"
 import { CodeMapMouseEventService, CursorType } from "../codeMap/codeMap.mouseEvent.service"
 import { Group, Mesh, PerspectiveCamera, Raycaster, Vector2, WebGLRenderer } from "three"
@@ -8,22 +7,20 @@ import * as Three from "three"
 import oc from "three-orbit-controls"
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls"
 import { ThreeOrbitControlsService } from "../codeMap/threeViewer/threeOrbitControlsService"
+import { EventEmitter } from "tsee"
+import { Inject, Injectable } from "@angular/core"
+import { ThreeOrbitControlsServiceToken } from "../../services/ajs-upgraded-providers"
 
-export interface ViewCubeEventPropagationSubscriber {
-	onViewCubeEventPropagation(eventType: string, event: MouseEvent)
+type ViewCubeEvents = {
+	viewCubeEventPropagation: (data: { event: MouseEvent; type: string }) => void
+	viewCubeHoveredEvent: (data: { cube: Mesh }) => void
+	viewCubeUnHoveredEvent: (data: undefined) => void
+	viewCubeClicked: (data: { cube: Mesh }) => void
 }
 
-export interface ViewCubeEventSubscriber {
-	onCubeHovered(cube: Mesh)
-	onCubeUnhovered()
-	onCubeClicked(cube: Mesh)
-}
-
+@Injectable({ providedIn: "root" })
 export class ViewCubeMouseEventsService {
-	private static VIEW_CUBE_EVENT_PROPAGATION_EVENT_NAME = "view-cube-event-propagation"
-	private static VIEW_CUBE_HOVER_EVENT_NAME = "view-cube-hover-event"
-	private static VIEW_CUBE_UNHOVER_EVENT_NAME = "view-cube-unhover-event"
-	private static VIEW_CUBE_CLICK_EVENT_NAME = "view-cube-click-event"
+	private eventEmitter = new EventEmitter<ViewCubeEvents>()
 
 	private cubeGroup: Group
 	private camera: PerspectiveCamera
@@ -32,9 +29,7 @@ export class ViewCubeMouseEventsService {
 	private controls: OrbitControls
 	private isDragging = false
 
-	constructor(private $rootScope: IRootScopeService, private threeOrbitControlsService: ThreeOrbitControlsService) {
-		"ngInject"
-	}
+	constructor(@Inject(ThreeOrbitControlsServiceToken) private threeOrbitControlsService: ThreeOrbitControlsService) {}
 
 	init(cubeGroup: Group, camera: PerspectiveCamera, renderer: WebGLRenderer) {
 		this.cubeGroup = cubeGroup
@@ -85,9 +80,9 @@ export class ViewCubeMouseEventsService {
 		this.controls.enableRotate = value
 	}
 
-	private checkMouseIntersection(event: MouseEvent, mouseAction: string) {
+	private checkMouseIntersection(event: MouseEvent, type: string) {
 		if (!this.getCubeIntersectedByMouse(event)) {
-			this.triggerViewCubeEventPropagation(mouseAction, event)
+			this.eventEmitter.emit("viewCubeEventPropagation", { type, event })
 		}
 	}
 
@@ -138,7 +133,7 @@ export class ViewCubeMouseEventsService {
 			if (this.currentlyHovered) {
 				this.triggerViewCubeUnhoverEvent()
 			}
-			this.triggerViewCubeEventPropagation("mousemove", event)
+			this.eventEmitter.emit("viewCubeEventPropagation", { type: "mousemove", event })
 		}
 	}
 
@@ -146,55 +141,27 @@ export class ViewCubeMouseEventsService {
 		this.isDragging = false
 		const cube = this.getCubeIntersectedByMouse(event)
 		if (cube) {
-			this.triggerViewCubeClickEvent(cube)
+			this.eventEmitter.emit("viewCubeClicked", { cube })
 		} else {
-			this.triggerViewCubeEventPropagation("mouseup", event)
+			this.eventEmitter.emit("viewCubeEventPropagation", { type: "mouseup", event })
 		}
-	}
-
-	private triggerViewCubeEventPropagation(type: string, event: MouseEvent) {
-		this.$rootScope.$broadcast(ViewCubeMouseEventsService.VIEW_CUBE_EVENT_PROPAGATION_EVENT_NAME, {
-			e: event,
-			type
-		})
 	}
 
 	private triggerViewCubeHoverEvent(cube: Mesh) {
 		this.currentlyHovered = cube
 		CodeMapMouseEventService.changeCursorIndicator(CursorType.Pointer)
-		this.$rootScope.$broadcast(ViewCubeMouseEventsService.VIEW_CUBE_HOVER_EVENT_NAME, { cube })
+		this.eventEmitter.emit("viewCubeHoveredEvent", { cube })
 	}
 
 	private triggerViewCubeUnhoverEvent() {
 		this.currentlyHovered = null
 		CodeMapMouseEventService.changeCursorIndicator(CursorType.Default)
-		this.$rootScope.$broadcast(ViewCubeMouseEventsService.VIEW_CUBE_UNHOVER_EVENT_NAME)
+		this.eventEmitter.emit("viewCubeUnHoveredEvent", undefined)
 	}
 
-	private triggerViewCubeClickEvent(cube: Mesh) {
-		this.$rootScope.$broadcast(ViewCubeMouseEventsService.VIEW_CUBE_CLICK_EVENT_NAME, { cube })
-	}
-
-	static subscribeToEventPropagation($rootScope: IRootScopeService, subscriber: ViewCubeEventPropagationSubscriber) {
-		$rootScope.$on(
-			ViewCubeMouseEventsService.VIEW_CUBE_EVENT_PROPAGATION_EVENT_NAME,
-			(_event_, parameters: { type: string; e: MouseEvent }) => {
-				subscriber.onViewCubeEventPropagation(parameters.type, parameters.e)
-			}
-		)
-	}
-
-	static subscribeToViewCubeMouseEvents($rootScope: IRootScopeService, subscriber: ViewCubeEventSubscriber) {
-		$rootScope.$on(ViewCubeMouseEventsService.VIEW_CUBE_HOVER_EVENT_NAME, (_event_, parameters: { cube: Mesh }) => {
-			subscriber.onCubeHovered(parameters.cube)
-		})
-
-		$rootScope.$on(ViewCubeMouseEventsService.VIEW_CUBE_UNHOVER_EVENT_NAME, () => {
-			subscriber.onCubeUnhovered()
-		})
-
-		$rootScope.$on(ViewCubeMouseEventsService.VIEW_CUBE_CLICK_EVENT_NAME, (_event_, parameters: { cube: Mesh }) => {
-			subscriber.onCubeClicked(parameters.cube)
+	subscribe<Key extends keyof ViewCubeEvents>(key: Key, callback: ViewCubeEvents[Key]) {
+		this.eventEmitter.on(key, data => {
+			callback(data)
 		})
 	}
 }