+++ visualization/mocks/ng.mockhelper.ts
@@ -3,7 +3,6 @@ import { IdToBuildingService } from "../app/codeCharta/services/idToBuilding/idT
 import { Store } from "../app/codeCharta/state/store/store"
 
 export const NGMock: IAngularStatic = angular
-export const NG = angular
 
 export function instantiateModule(id: string) {
 	NGMock.mock.module(id)
@@ -11,12 +10,20 @@ export function instantiateModule(id: string) {
 	angular.mock.module(($provide: ng.auto.IProvideService) => {
 		$provide.value("idToBuilding", idToBuildingService)
 	})
+	angular.mock.module(($provide: ng.auto.IProvideService) => {
+		$provide.value("viewCubeMouseEvents", {})
+	})
 }
 
 export function getService<T>(id: string): T {
 	// eslint-disable-next-line prefer-const
 	let service: T = null
 	eval(`NGMock.mock.inject(function(_${id}_) { service = _${id}_; });`)
-	if (id === "storeService") Store["initialize"]() // reset shared store
+	if (id === "storeService") {
+		// reset shared store
+		Store["initialize"]()
+	}
 	return service
 }
+
+export { default as NG } from "angular"