+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 -   Fix missing color pickers within edge metric options [#2993](https://github.com/MaibornWolff/codecharta/pull/2993)
 -   Fix labels being displayed at an incorrect height and lines being disconnected from labels [#2990](https://github.com/MaibornWolff/codecharta/pull/2991)
 
+### Chore 👨‍💻 👩‍💻
+
+-   Migrate view cube component and its service to Angular [#2998](https://github.com/MaibornWolff/codecharta/pull/2998)
+
 ## [1.103.6] - 2022-08-17
 
 ### Fixed 🐞